misc: Refactor structure
